requirements.txt
Flask
numpy
matplotlib
Pillow
scikit-learn
